+++ kaiax/valset/impl/init.go
@@ -25,7 +25,6 @@ import (
 	"github.com/kaiachain/kaia/common"
 	"github.com/kaiachain/kaia/consensus"
 	"github.com/kaiachain/kaia/kaiax/gov"
-	"github.com/kaiachain/kaia/kaiax/gov/headergov"
 	"github.com/kaiachain/kaia/kaiax/staking"
 	"github.com/kaiachain/kaia/kaiax/valset"
 	"github.com/kaiachain/kaia/log"
@@ -162,141 +161,4 @@ func (v *ValsetModule) migrate() {
 
 	// Now the migration is complete.
 	writeLowestScannedVoteNum(v.ChainKv, 0)
-}
-
-// getCouncilFromIstanbulSnapshot re-generates the council at the given targetNum.
-// Returns the council at targetNum, the nearest snapshot number, and error if any.
-//
-// The council is calculated from the nearest istanbul snapshot (at snapshotNum)
-// plus the validator votes in the range [snapshotNum+1, targetNum-1]. Note that
-// snapshot at snapshotNum already reflects the validator vote at snapshotNum,
-// so we apply the votes starting from snapshotNum+1.
-//
-// If write is true, ValidatorVoteBlockNums and Council in the extended range
-// [snapshotNum+1, targetNum] are written to the database. Note that this time
-// the targetNum is included in the range for completeness. This property is
-// useful for snapshot interval-wise migration.
-func (v *ValsetModule) getCouncilFromIstanbulSnapshot(targetNum uint64, write bool) (*valset.AddressSet, uint64, error) {
-	if targetNum == 0 {
-		council, err := v.getCouncilGenesis()
-		return council, 0, err
-	}
-	// Load council at the nearest istanbul snapshot. This is the result
-	// applying the votes up to the snapshotNum.
-	snapshotNum := roundDown(targetNum-1, istanbulCheckpointInterval)
-	header := v.Chain.GetHeaderByNumber(snapshotNum)
-	if header == nil {
-		return nil, 0, errNoHeader
-	}
-	council := valset.NewAddressSet(ReadIstanbulSnapshot(v.ChainKv, header.Hash()))
-	if council.Len() == 0 {
-		return nil, 0, ErrNoIstanbulSnapshot(snapshotNum)
-	}
-
-	// Apply the votes in the interval [snapshotNum+1, targetNum-1].
-	for n := snapshotNum + 1; n < targetNum; n++ {
-		if err := v.replayBlock(council, n, write); err != nil {
-			return nil, 0, err
-		}
-	}
-	// Apply the vote at targetNum to write to database, but do not return the modified council.
-	if write {
-		// Apply the vote at targetNum and write to database, but do not affect the returning council.
-		if err := v.replayBlock(council.Copy(), targetNum, true); err != nil {
-			return nil, 0, err
-		}
-	}
-	return council, snapshotNum, nil
-}
-
-func (v *ValsetModule) replayBlock(council *valset.AddressSet, num uint64, write bool) error {
-	header := v.Chain.GetHeaderByNumber(num)
-	if header == nil {
-		return errNoHeader
-	}
-	governingNode := v.GovModule.EffectiveParamSet(num).GoverningNode
-	if applyVote(header, council, governingNode) && write {
-		insertValidatorVoteBlockNums(v.ChainKv, num)
-		writeCouncil(v.ChainKv, num, council.List())
-	}
-	return nil
-}
-
-// applyVote modifies the given council *in-place* by the validator vote in the given header.
-// governingNode, if specified, is not affected by the vote.
-// Returns true if the council is modified, false otherwise.
-func applyVote(header *types.Header, council *valset.AddressSet, governingNode common.Address) bool {
-	voteKey, addresses, ok := parseValidatorVote(header)
-	if !ok {
-		return false
-	}
-
-	originalSize := council.Len()
-	for _, address := range addresses {
-		if address == governingNode {
-			continue
-		}
-		switch voteKey {
-		case gov.AddValidator:
-			if !council.Contains(address) {
-				council.Add(address)
-			}
-		case gov.RemoveValidator:
-			if council.Contains(address) {
-				council.Remove(address)
-			}
-		}
-	}
-	return originalSize != council.Len()
-}
-
-func parseValidatorVote(header *types.Header) (gov.ValidatorParamName, []common.Address, bool) {
-	// Check that a vote exists and is a validator vote.
-	voteBytes := headergov.VoteBytes(header.Vote)
-	if len(voteBytes) == 0 {
-		return "", nil, false
-	}
-	vote, err := voteBytes.ToVoteData()
-	if err != nil {
-		return "", nil, false
-	}
-	voteKey := gov.ValidatorParamName(vote.Name())
-	_, isValidatorVote := gov.ValidatorParams[voteKey]
-	if !isValidatorVote {
-		return "", nil, false
-	}
-
-	// Type cast the vote value. It can be a single address or a list of addresses.
-	var addresses []common.Address
-	switch voteValue := vote.Value().(type) {
-	case common.Address:
-		addresses = []common.Address{voteValue}
-	case []common.Address:
-		addresses = voteValue
-	default:
-		return "", nil, false
-	}
-
-	return voteKey, addresses, true
-}
-
-func roundDown(n, p uint64) uint64 {
-	return n - (n % p)
-}
-
-// TODO-kaiax: move the feature into staking_info.go
-func collectStakingAmounts(nodes []common.Address, si *staking.StakingInfo) map[common.Address]float64 {
-	cns := si.ConsolidatedNodes()
-	stakingAmounts := make(map[common.Address]float64, len(nodes))
-	for _, node := range nodes {
-		stakingAmounts[node] = 0
-	}
-	for _, cn := range cns {
-		for _, node := range cn.NodeIds {
-			if _, ok := stakingAmounts[node]; ok {
-				stakingAmounts[node] = float64(cn.StakingAmount)
-			}
-		}
-	}
-	return stakingAmounts
 }
